pyproject.toml
[project]
name = "concrete-ml-playground"
version = "0.1.0"
description = ""
authors = [
    {name = "JulianFP",email = "julian@partanengroup.de"}
]
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "concrete-ml (>=1.9.0,<2.0.0)",
    "click (>=8.2.1,<9.0.0)",
    "datasets (>=4.0.0,<5.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "pydantic (>=2.11.7,<3.0.0)"
]

[tool.poetry]
packages = [{include = "concrete_ml_playground", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
